Implementación de autorizaciones para funciones de tareas

+++ backend/controllers/tareasController.js
@@ -2,7 +2,7 @@ const asyncHandler = require('express-async-handler')
 const Tarea = require('../model/tareasModel')
 
 const getTareas = asyncHandler(async (req, res) => {
-    const tareas = await Tarea.find()
+    const tareas = await Tarea.find({ user: req.user.id })
     res.status(200).json(tareas)
 })
 
@@ -14,6 +14,7 @@ const setTareas = asyncHandler(async (req, res) => {
 
     const tarea = await Tarea.create({
         texto: req.body.texto,
+        user: req.user.id
     })
 
     res.status(201).json({ tarea })
@@ -23,14 +24,21 @@ const updateTareas = asyncHandler(async (req, res) => {
 
     const tarea = await Tarea.findById(req.params.id)
 
+    //verificamos que la tarea existe
     if (!tarea) {
         res.status(404)
         throw new Error("tarea no encontrada")
     }
 
-    const tareaUpadated = await Tarea.findByIdAndUpdate(req.params.id, req.body, { new: true })
+    //verificar que la tarea pertenezca al usuario del token que la quiere modificar
+    if (tarea.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('Acceso no autorizado')
+    } else {
+        const tareaUpadated = await Tarea.findByIdAndUpdate(req.params.id, req.body, { new: true })
 
-    res.status(200).json(tareaUpadated)
+        res.status(200).json(tareaUpadated)
+    }
 })
 
 const deleteTareas = asyncHandler(async (req, res) => {
@@ -42,13 +50,19 @@ const deleteTareas = asyncHandler(async (req, res) => {
         throw new Error("tarea no encontrada")
     }
 
-    await Tarea.deleteOne(tarea)
+    //verificar que la tarea pertenezca al usuario del token que la quiere modificar
+    if (tarea.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('Acceso no autorizado')
+    } else {
+        await Tarea.deleteOne(tarea)
 
-    //Este es otro metodo que podríamos utilizar pero no esta eficiente dado que tendríamos que hacer
-    //una búsqueda en la base de datos para eliminarlo cosa que ya realizamos anterior mente
-    //await Tarea.findByIdAndDelete(req.params.id)
+        //Este es otro metodo que podríamos utilizar pero no esta eficiente dado que tendríamos que hacer
+        //una búsqueda en la base de datos para eliminarlo cosa que ya realizamos anterior mente
+        //await Tarea.findByIdAndDelete(req.params.id)
 
-    res.status(200).json({ id: tarea.id })
+        res.status(200).json({ id: tarea.id })
+    }
 })
 
 module.exports = {
